+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 
+- [#965](https://github.com/hyperf/hyperf/pull/965) Added Redis Lua Module.
 - [#1023](https://github.com/hyperf/hyperf/pull/1023) Added CUSTOM_MODE to hyperf/metric prometheus driver.
 
 ## Fixed
@@ -17,6 +18,7 @@
 
 - [#1014](https://github.com/hyperf/hyperf/pull/1014) Optimized `Command:: execute` return value does not support null.
 - [#1022](https://github.com/hyperf/hyperf/pull/1022) Provided cleaner connection pool error message without implementation details.
+- [#1034](https://github.com/hyperf/hyperf/pull/1034) The property `arguments` of `Hyperf\Amqp\Builder\Builder` not only support array.
 
 # v1.1.7 - 2019-11-21